+++ bot/libs/utils/pages/__init__.py
@@ -1,5 +1,4 @@
 from .paginator import KumikoPages as KumikoPages
 from .sources import (
-    EmbedListSource as EmbedListSource,
     SimplePageSource as SimplePageSource,
 )
