+++ bot/libs/utils/pages/paginator.py
@@ -6,7 +6,22 @@ import discord
 from discord.ext import menus
 from discord.ext.commands import Context
 
-from .modals import NumberedPageModal
+
+class NumberedPageModal(discord.ui.Modal, title="Go to page"):
+    page = discord.ui.TextInput(
+        label="Page", placeholder="Enter a number", min_length=1
+    )
+
+    def __init__(self, max_pages: Optional[int]) -> None:
+        super().__init__()
+        if max_pages is not None:
+            as_string = str(max_pages)
+            self.page.placeholder = f"Enter a number between 1 and {as_string}"
+            self.page.max_length = len(as_string)
+
+    async def on_submit(self, interaction: discord.Interaction) -> None:
+        self.interaction = interaction
+        self.stop()
 
 
 # This is originally from RoboDanny's Paginator class (RoboPages)
@@ -148,7 +163,10 @@ class KumikoPages(discord.ui.View):
     async def start(
         self, *, content: Optional[str] = None, ephemeral: bool = False
     ) -> None:
-        if self.check_embeds and not self.ctx.channel.permissions_for(self.ctx.me).embed_links:  # type: ignore
+        if (
+            self.check_embeds
+            and not self.ctx.channel.permissions_for(self.ctx.me).embed_links
+        ):  # type: ignore
             await self.ctx.send(
                 "Bot does not have embed links permission in this channel.",
                 ephemeral=True,
