+++ old_code/libs/ui/config/log_views.py
