+++ old_code/libs/ui/search/structs.py
